requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
textblob
python-dotenv
sentence-transformers